+++ packages/dmn-js-drd/lib/features/drill-down/DrillDown.js
@@ -35,12 +35,12 @@ export default class DrillDown {
 
   constructor(injector, eventBus, overlays, config) {
     this._injector = injector;
+    this._eventBus = eventBus;
     this._overlays = overlays;
 
     this._config = config || { enabled: true };
-    this._eventBus = eventBus;
 
-    this._eventBus.on([ 'drdElement.added', 'shape.added' ], ({ element }) => {
+    eventBus.on([ 'shape.added' ], ({ element }) => {
 
       for (let i = 0; i < PROVIDERS.length; i++) {
 
@@ -51,13 +51,6 @@ export default class DrillDown {
         if (editable) {
           this.addOverlay(element, className);
         }
-      }
-    });
-
-    this._eventBus.on('drillDown.click', ({ decision, parent }) => {
-      var view = parent.getView(decision);
-      if (view) {
-        parent.open(view);
       }
     });
   }
@@ -93,26 +86,57 @@ export default class DrillDown {
   }
 
   /**
-   *
    * @param {Object} element
    * @param {Object} overlay
    * @param {string} id
    */
   bindEventListener(element, overlay, id) {
     const overlays = this._overlays,
-          injector = this._injector;
-
-    var parent = injector.get('_parent', false);
-
-    if (!parent) {
-      return;
-    }
+          eventBus = this._eventBus;
 
     const overlaysRoot = overlays._overlayRoot;
-    const { _eventBus } = this;
+
     domDelegate.bind(overlaysRoot, '[data-overlay-id="' + id + '"]', 'click', () => {
-      _eventBus.fire('drillDown.click', { decision: element.businessObject, parent });
+
+      const triggerDefault = eventBus.fire('drillDown.click', {
+        element
+      });
+
+      if (triggerDefault === false) {
+        return;
+      }
+
+      this.drillDown(element);
     });
+  }
+
+
+  /**
+   * Drill down into the specific element.
+   *
+   * @param  {djs.model.Base} element
+   *
+   * @return {Boolean} whether drill down was executed
+   */
+  drillDown(element) {
+
+    const parent = this._injector.get('_parent', false);
+
+    // no parent; skip drill down
+    if (!parent) {
+      return false;
+    }
+
+    const view = parent.getView(element.businessObject);
+
+    // no view to drill down to
+    if (!view) {
+      return false;
+    }
+
+    parent.open(view);
+
+    return true;
   }
 
 }