+++ packages/dmn-js/CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [dmn-js](https://github.com/bpmn-io/dmn-js) are documente
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT(drd)`: add ability to intercept drill-down via event listener ([#353](https://github.com/bpmn-io/dmn-js/issues/353))
+
 ## 4.2.1
 
 * `FIX(decision-table)`: close input editor on `ENTER`